refactor: move parsers, clients to apps folder
