+++ apps/bot/src/apps/deal/abstracts/deal.parser.ts
