+++ apps/bot/src/apps/deal/deal.error-message.ts
