+++ apps/bot/src/apps/offer/offer.client.ts
