+++ apps/bot/src/apps/offer/offer.error-message.ts
