+++ apps/bot/src/apps/offer/offer.providers.ts
@@ -1,0 +1,4 @@
+import { OfferClient } from './offer.client';
+import { OfferParser } from './offer.parser';
+
+export const BOT_OFFER_PROVIDERS = [OfferClient, OfferParser];
