+++ apps/bot/src/apps/offer/offer.types.ts
