--- apps/bot/src/clients/clients.ts
@@ -1,3 +1,0 @@
-import { UserClient } from './user/user.client';
-
-export const CLIENTS = [UserClient];
