--- apps/bot/src/clients/user-image/user-image.client.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { CommandBus } from '@nestjs/cqrs';
-
-@Injectable()
-export class UserImageClient {
-  constructor(private readonly commandBus: CommandBus) {}
-
-  // TODO
-}
