--- apps/bot/src/pipes/parsers.ts
@@ -1,3 +1,0 @@
-import { OfferParser } from './deal/offer/offer.parser';
-
-export const PARSERS = [OfferParser];
