--- apps/bot/src/pipes/user/user.types.ts
@@ -1,9 +1,0 @@
-import { Message } from 'discord.js';
-
-export interface UserWithMessageType {
-  user: {
-    id: string;
-  };
-
-  message: Message;
-}
